Change: added getFrom()/getTo() operations
Each change is associated with a source FileInRevision describing which
revision of which file a change is based upon, and a target
FileInRevision describing which FileInRevision a change was integrated
into. Now a Change can be queried for this information by using
getFrom() and getTo() operations. Due to a name conflict, the operations
getFrom() and getTo() of class TextualChangeHunk have been renamed to
getFromFragment() and getToFragment(), respectively.