svn: use correct revision in SvnFragmentTracer.traceFile()
Until now, a LocalRevision was used to describe the FileInRevision of
a traced fragment. This is not correct, however, as the code analyses
the repository revisions only, so uncommitted renames within the
working copy are not detected anyway. So this changeset makes
traceFile() return a RepoRevision associated with the last revision
that touches the given file path.