add "invisible" attribute to changes/stops and commits/tours
When analysing changes, it may happen that not all file changes belong
to the same commit/tour. If there are "holes" in a file's history, this
may lead to errors, especially when computing differences between
commits. This changeset allows to attach an "invisible" attribute to
changes/stops and commits/tours in order to include them in difference
creation but to exclude them from being displayed to the user. A new
abstraction IReviewElement has been added to the changestructure package
to be able to handle stops and tours uniformly.