changestructure: added change history types
The new types RepositoryChangeHistory and FileChangeHistory make it easy
to manage all changes in a file or repository, respectively, with regard
to a given revision range. Contrary to the FileDiff type, a
FileChangeHistory does not merge the Hunks. Consequently, it is possible
to generate different FileDiffs on-the-fly, providing different revision
ranges. The RepositoryChangeHistory type manages FileChangeHistories for
a given file, abstracting from path changes (renames).