svn: handle checkouts of multiple branches
If the user checks out multiple branches of the same repository, the
commit detection code is confused because the same commit is associated
with all checkouts, even if it refers only files from a single checkout.
This is because the log action does not filter wrt. the repository URLs
of the working directories. This changeset makes the log action consider
the working directory's repository URL.