changestructure, svn: added ability to get a file's whole contents
For visualizing more context of a diff, we need lines that are not part
of any fragment because they have not been changed. In order to do this,
each FileInRevision can have a IContentSource attached which can be
asked to retrieve the file's contents. As the file contents are loaded
anyway by the svn plug-in for each file while the commits are analysed,
a cache (SvnFileCache) has been introduced to cache these file contents
in order to avoid to query the repository unnecessarily.