tourrestructuring: fix merging of stops
The current algorithm does not merge all possible stops. Consider the
situation where in tour T1 a stop S1 creates a file F, and in tour T2
three stops S2, S3, and S4 modify that file F. While analysing, S1 is
determined to be mergeable with S2. The algorithm stores the merge
result S2' in T2/S2 (overwriting the old stop S2) and removes the stop
S1 from Tour T1. But the algorithm never considers whether the new
merged stop S2' can be merged with S3 or S4, because they are in the
same tour T2.

This changeset fixes this problem by also checking whether stops in the
same tour can be merged (preventing a stop from being merged with
itself).