Fix traversing non-linear history
The diff calculation did not work properly in the case of file history
branches, as only one descendant was chosen indeterministically. This
changeset fixes the buggy behaviour by two steps. The first step is to
make FileHistoryGraph manage the *whole* relevant SVN history. That
means that not only copies and deletions of files are tracked, but also
of directories (which includes tracking of child/parent relationships)
as well as of changes and the differences associated with them (which
are stored in FileDiff objects). This makes the formerly introduced
classes RepositoryChangeHistory and FileChangeHistory superfluous.
In order to make the core code access the information in the history
graph, a suitable interface has been extracted, being called
FileHistoryGraph, while the implementation is now called
SvnFileHistoryGraph.
The interfaces IChangeData and IFragmentTracer have also been adapted to
the new interface FileHistoryGraph, as well as the code found in
ToursInReview and CombinedDiffStopViewer.

The second step is to correct fragment tracing by introducing a new
operation traceFragment() in the class FileDiff and by updating the
SvnFragmentTracer to use it. The old approach was not correct in all
cases, leading to wrong results in some situations. Consequently, the
interfaces of FileDiff and FragmentList have been changed and/or extended
a bit.