svn: copy handling improved
If a file starts its life with a change in a copied directory, it is now
associated with the correct copy source, namely with a file in the copy
source of its parent directory having the same name. Likewise,
intermediate parent directories are also associated with the correct copy
sources if their (direct or indirect) parent is a copy.