svn: cache inconsistency handled gracefully
In commit 98a125ff747d ("svn: copy handling improved"), the member "paths"
of class CachedLogEntry changed its type from LinkedHashMap to TreeMap. If
CachedLog.readCacheFromFile() attempts to load an old cache, this fails
with a ClassCastException ("cannot assign instance of LinkedHashMap to field
paths of type SortedMap in instance of CachedLogEntry"). However, the
ClassCastException is not caught, causing the exception to abort the
initialization of a static attribute and therefore to abort the class
loading process. This changeset fixes this problem by explicitly catching
a ClassCastException.