changestructure: Stop manages Hunks instead of Fragments
Until now, a Stop used a list of Fragments in order to describe its
change history. This was cumbersome, as the relation between source
and target fragment was not explicit, and clients had to pair the
fragments on their own. This changeset makes the relation explicit
and adjusts the Stop API accordingly. On the one hand, the construction
of Stops has been simplified by accepting a TextualChangeHunk or a
BinaryHunk object. On the other hand, getHistory() returns
(source,target) pairs of FileInRevision objects, and getContentFor()
returns a list of Hunks instead of a list of Fragments. This makes it
easier to work with Stops.

As a side effect, Hunks are now ordered according to their source
Fragments. This is needed by Stop.merge() in order to keep the history
sorted.