changestructure: Fragments now keep track of their origins
During its lifetime, a Fragment may be transformed multiple times until
it is finally placed e.g. in a FileDiff between two revisions of a file.
Such transformations may be simple moves (forwards and backwards), but
also enlargements due to overlaps or adjoint situations. Those
enlargements typically originate from Fragments referring to different
revisions of the same (logical) file, e.g. when combining adjacent
changes to a file made in subsequent revisions. This changeset makes
Fragments keep track of their origin(s) such that it is clear where a
Fragment originally came from (think of moving fragments) and of which
original parts it is comprised (think of combining fragments).

In order to achieve this, the Fragment class has been split into two
parts: the original Fragment class which denotes Fragments of a single
FileInRevision, and the CombinedFragment class which denotes Fragments
that result from combining multiple Fragments belonging to potentially
different FileInRevision objects. A greatest common denominator, the
AbstractFragment class, holds both parts together and provides most of
the original Fragment API. In addition, an AbstractFragment now offers
the new operations getOrigins() and containsChangeInOneOf(), which
ask for or deal with the origins of an AbstractFragment, respectively.