changestructure: FileDiff knows about (from,to) FileInRevisions
Until now, merging new hunks into a FileDiff led to manipulating hunks
using the Fragment and FragmentList APIs which are agnostic of
FileInRevisions. This changeset puts the information about source and
target FileInRevisions into FileDiff, which makes the Hunks contained
in a FileDiff always hold Fragments with the correct source and target
FileInRevisions.