changestructure: FileDiff: deal with inline diffs
Since in-line diffs were introduced, the FileDiff algorithms have not
worked correctly in all cases. The reason for this is that the original
algorithms were designed for complete-line diffs and that the
implemented workaround of making in-line diffs full lines was not
accurate in some situations, e.g. when the contents of a line were
deleted but the whole line in itself remained. This changeset adapted
the FileDiff algorithms to be able to deal with in-line diffs. The new
class Delta encapsulates a position difference consisting of a line
and a column offset.

A necessary side effect is that all position ranges are now half-open
intervals, i.e. the second position is exclusive. This was needed in
order to properly calculate with positions and deltas.