changestructure/changesources.svn: FileHistoryGraph generalized
Most of the code in FileHistoryGraph and related classes can be reused
with other VCSs. This changeset splits the implementation into a general
and a Subversion specific part and moves the general part to the core
project. A side effect of the generalization is the support for multiple
ancestors of a node (think e.g. of merges in hg or git). Another change
is the addition of UnknownRevision for denoting an unknown revision;
it is used for superficial ancestor nodes of added files as revision
number arithmetics is not possible with every VCS.

Note that the CombinedDiffStopViewer has not been generalized to
handle multiple ancestors, so only the diff along the first merge
parent(s) is displayed.