changesources.svn: consider changes in the working copy
This changeset makes the Subversion change source aware of local
modifications, such that the correct most recent files and fragments
are used when the working copy is not clean. This made some rewrite
of FileHistoryGraph and associated classes necessary. The first one is
the introduction of yet another revision abstraction ISvnRevision for
SvnRevision and a new subclass WorkingCopyRevision in order to be able
to differentiate between repository revisions and the working copy
revision in the Subversion change source.

The second change deals with the fact that in a working copy, a path
may have been added and deleted simultaneously, e.g. if you copy a
directory followed by renaming some file in it. So FileHistoryNodes may
change their "deleted" state. Because of that, ExistingFileHistoryNode
and NonExistingFileHistoryNode have been removed in favour of a
changeable boolean flag "isDeleted" in FileHistoryNode (which is not
abstract anymore). Another consequence is that each CachedLogEntryPath
now knows its ancestor revision, even if no copy has been done. This is
important for local changes in the working copy because all these
changes belong to the same revision, so calculating an ancestor revision
by decrementing the descendant's revision does not work.

The third change extends the Repository interface by a new operation
fromAbsolutePathInWc(). This operation is necessary because some of the
working copy status information SVNKit provides returns absolute file
system paths.

Finally, some corner case bugs in copy handling have been fixed.