ui: always open resources in an internal editor
Opening a resource in an external editor is bad for two reasons. First,
if the resource does not exist, the external editor is called anyway,
typically producing an error message (this depends on the external
editor). Second, if the resource exists, the external editor might be
not an editor at all. For example, batch scripts are executed (!)
instead of being displayed. This changeset prevents these problems by
calling openInternalEditorOnFileStore() instead of
openEditorOnFileStore().