changestructure/ui: separate repository and local changes
This changeset separates changes originating from commits in the
repository from changes found in the local working copy. A new
IFileHistoryGraph implementation, VirtualFileHistoryGraph, allows to
combine two different IFileHistoryGraphs into one. The IChangeSource and
IChangeData interfaces have been adapted to allow for this combination.
A side effect is the move of SvnFragmentTracer to the core package
(now called FragmentTracer there) because there is no Subversion
specific behaviour anymore in this class.

The local modifications are kept in a special tour, the local tour.
This is because local changes should not be merged with repository
changes in order to avoid mixing the author's changes to be reviewed by
the reviewer and the reviewer's code to be reviewed later by some other
reviewer. Because of this, this local tour is also never merged with
any other tour.

Note that the ToursInReview and Tour classes remain immutable in the
sense that local modifications do not have any impact on tour
restructuring, i.e. a Tour keeps the same Stops regardless of local
modifications. However, stops remember in addition a most recent
fragment/file considering local changes in order to have correct markers
even if the files are modified after the review has started. Note also
that the immutability leads to the situation that new descendants are
not shown when visiting a stop resulting from a repository change. Such
a situation happens when a file is locally copied and subsequently
changed. These changes are only visible in the local tour, but making
them trackable in the tour for the repository commit would require
adding new stops, which is currently prohibited.