CachedLogEntryPath: use last committed revision
When determining the ancestor revision for a non-copied path
with uncommitted modifications, the last committed revision is now used
instead of the revision the working copy has been updated to. This
allows later to merge file history graphs as the revision of the last
commit is equal to the ancestor revision of the locally modified path.