ReviewPlugin: update Stops and file history on local changes
The IResourceChangeListener implementation has been extended to check
for local changes which influence the local tour. This is a superset of
the changes used for telemetry, because it also includes additions,
removals, and replacements. If relevant local changes have been
detected, ToursInReview.createLocalTour() is called, which then leads
to recomputation of the local tour and the most recent fragments/files
of stops.

In order to speed up the detection of local changes, the whole working
copy is analyzed only once at review start. Later on, the paths of
changed resources are computed and passed to the IChangeSource which
can use them in order to only analyze those paths instead of the whole
working copy. The SvnChangeSource implements this optimization. The
ToursInReview class stores the local paths of the relevant changes
retrieved by using the new operation IChangeData.getLocalPaths(), such
that incremental changes can be dealt with properly.

For better understanding, IChangeSource.getChanges() has been renamed
to IChangeSource.getRepositoryChanges().