changestructure: Handling local changes moved to ChangeManager
The handling of local changes has been moved from ToursInReview to
ChangeManager. This has been done as monitoring local changes is not
only important while reviewing but also while fixing. In the future,
the ChangeManager could be initialized at plugin start and stay active
even if no review or fixing is being done.