model: PositionTransformer: choose path with minimal number of segments
When searching a resource for a short name path, the first matching one
was returned. This changeset analyzes all matching resources and returns
one having a path with a minimal number of segments. So if a/1.txt is
to be resolved and source/x/a/1.txt as well as source/a/1.txt both
match, than the shorter one is returned as the probability that it is
the correct path is higher.