changestructure: FileHistoryGraph: use the original edge's type when creating a new one
Although a check exists that guarantees that the edge is a NORMAL edge,
it more understandable to retrieve the original edge's type, as it
emphasizes the fact that the edge type be not changed. In addition,
it is safer under refactoring when this part of the code is copied or
moved to some other place where the condition is not fulfilled.