changestructure: FileHistoryGraph: added nodes of type UNCONFIRMED
Such nodes are created when it is unknown to which flow a node belongs,
e.g. because it is a copy source from an otherwise unknown revision.
(They were called "alpha ancestors" until now.) Unconfirmed nodes
cannot be confirmed later on, as it is assumed that revisions are
processed from older to newer ones.