svn: RelevantRevisionLookupHandler: improve handling of replaced files
Replaced nodes are a bit mysterious as they are part of two different
flows. A replaced node is both deleted (having a non-deleted parent
revision) and added or copied (having a non-deleted copy source).
This changeset improves the determination of the relevant revisions and
makes RelevantRevisionLookupHandler call the addReplacement() operation
on the IMutableFileHistoryGraph if a replacement has been detected.