changestructure: FileHistoryGraph tracks copies by special edges
Until now, copies were detected by comparing the associated paths.
This is not accurate, however, because a copy can have the same path
as its origin provided there was a deletion in between (this happens if
an old revision is resurrected after a deletion). Now, copies are
explicitly tracked using edges with the type COPY or COPY_DELETED.
While at this, the operation FileHistoryNode.isCopied() has been
renamed to isCopyTarget() and moved to IFileHistoryNode.