changestructure: FileHistoryGraph now creates "alpha" ancestors
Now, every flow roots in a so-called alpha ancestor, i.e. a node with an
unknown revision. This will finally help in combining different graphs.

Due to the omnipresent root node, the flow detection has changed: Now a
copy source is always associated with subsequent changes when copying
from an unknown revision. However, this behaviour change is mostly
irrelevant in practice as unknown copy sources are neither reviewed nor
do they take part in diff computations (with the exception of the
branching flow).