changestructure: FileHistoryGraph: reduce number of nodes, fix deletion
For file history graphs spanning a large range of revisions, many nodes
are created many of which are superfluous as they do not denote an
addition, a change or a deletion but are only created for a more
complete child/parent relationship. As this relationship is not really
necessary for the history tracing algorithms, this change reduces the
number of nodes to be created.

As a side effect, deletion of directories has been corrected such that
all children are marked as deleted, not only child nodes found in the
last commit before the deletion.

Note that avoiding the creation of child nodes when when copying a
directory makes it almost impossible to complete incomplete copy
operations when combining different file history graphs, as the
information about the copied nodes can span over multiple file history
graphs. As creating a branch is a relatively seldom operation, this
type of optimization is not done. This also allows to handle copies in a
special way if desired, even if no textual changes happen to exist.

While at this, the FileHistoryGraph class has been largely refactored in
order to make the code better to understand and easier to extend. A
side effect is that additions of new and changes of existing files are
now kept separate by using different operations (addAddition() and
addChange(), respectively).