changestructure: FileHistoryGraph: new node types ADDED and CHANGED
The node type NORMAL has been split into ADDED and CHANGED in order to
be able to distinguish between nodes starting a new flow and nodes
continuing an existing flow. This again led to some refactorings of the
FileHistoryGraph class.

Furthermore, some shortcomings of the previous commit have been
corrected. This allowed to restore some testcases (notably those
operating in a copy of a previously unknown directory) to their original
state.