model: each IRepository knows "its" file history graph
Instead of having one big file history graph for all repositories, the
idea is to have one per repository. This is reasonable as it expresses
better the locality of paths wrt. repositories and it simplifies the
implementation of FileHistoryGraph.

This commit only implements the basic infrastructure, there is still one
global file history graph in use (managed by ChangeManager). The
integration of per-repository file history graphs will be done in later
commits.