model: FileHistoryGraph: completion of copy source nodes improved
Formerly, the algorithm worked recursively through the tree which did
not work for long chains due to stack overflow issues. Now, chains are
processed iteratively, and recursive invocations are only done at
branching points.

Moreover, if a node has already been completed, it is not necessary to
do it a second time. So, this changeset makes nodes remember whether
they know all their children, which makes processing copy operations
faster as it is only necessary to go back to the last branching point.