svn: loading cache is now done by a background job
If the cache is large (e.g. more than 50000 commits) and if operations
referring to many nodes (copying/moving or deleting top-level folders),
processing the commits can take much time (> 10 minutes). This commit
makes this a background operation such that if the developer does not
try to review directly after starting Eclipse, the chance is high that
she does not notice the cache load time at all.