svn: implemented one cache per working copy
This prevents the whole cache to be written each time any one of the
repositories receives new commits. In addition, it allows to share the
cache with other developers (provided the working copy roots are equal).

Note that switching a working copy to another branch is supported in the
sense that the cache name is based on the remote repository URL such
that the switched checkout will have its own SVN cache. However,
switching a working copy while Eclipse is running is not supported as
the working copies are analyzed only at plugin initialization time.