plugin: handle project additions and removals while running
Until now, the set of projects was determined when Eclipse was started.
Now, adding/opening and removing/closing projects is handled properly.
To achieve this, the ReviewPlugin now manages a set of known project
roots and informs the change source about them by using the new
operations IChangeSource.addProject() and IChangeSource.removeProject(),
respectively. This requires the ResourceChangeListener to be installed
at plug-in initialization time rather than when starting a review or
fixing session.

Note that in the long run, these changes will enable us to track local
changes even if no review is currently active, thereby reducing review
startup time considerably. This will be addressed in a separate commit.