telemetry: use a NoTelemetry object as initial value
If the CoRT configuration is missing or invalid, it is possible that the
TelemetryConfigurator is not executed, such that Telemetry.instance
remains null. This is bad because later NullPointerExceptions are
produced when the TelemetryEventBuilder is used to log telemetry events.
This commit sets Telemetry.instance initially to an object of type
NoTelemetry().