build: removed Eclipse .checkstyle files
These files are to be generated by the m2e-code-quality connector.
This has the advantage that the generated (absolute) path to the
Checkstyle configuration file is always valid and independent of a
specific placement of the checkout, thereby making the Checkstyle
configuration relocatable and the build configuration more portable.

Note that to make Checkstyle work in Eclipse, you need to install the
above mentioned m2e-code-quality connector.