model: don't consider local changes when creating stops
When stops are created in ToursInReview.toSliceFragment(), it is
important to use the file history graph of the underlying repository and
not the one of the working copy, thereby ignoring local modifications.
This is because the traced file or fragment is stored in the stop's
*original* most recent file or fragment, respectively, and serve as a
start for tracing local changes. If this fragment or file is already
associated with the local revision, no further changes will ever be
detected as the local revision has no descendants and tracing will not
yield any new results.

In order to eventually consider local modifications also for new stops,
the IChangeManagerListener.localChangeInfoUpdated() method calls the
operation updateMostRecentFragmentsWithLocalChanges() which then makes
all stops to retrace their fragments or files, respectively, using the
virtual file history graph which includes local changes.