model: VirtualFileHistoryGraph can bridge revision gaps
Until now, the VirtualFileHistoryGraph relied upon the fact that flows
to be connected between a remote and a local file history graph have
some revision in common. This commit removes this requirement: Now any
gap between the latest revision in the remote file history graph and
the earliest non-alpha revision in the local file history graph is
connected by an appropriate edge. To make this possible, the operation
findAncestorFor() has been pulled up to IFileHistoryGraph.