svn: fix tracing of local changes in files serving as copy sources
As the VirtualFileHistoryGraph can bridge revision gaps, we are able to
use the revision of the working copy instead of the committed revision
when processing local changes. This is necessary for the following
scenario:

(1) let R1 be the last committed revision of file F
(2) let R2 be some revision where F@R1 is copied to F'@R2 (R1 > R2)
(3) let F be changed locally

The old behaviour associated the local change with R1 as this is the
committed revision of F returned by Subversion. So we had:
  R1 < R2
  R1 < $
  R2 </ $
However, for the file history graph, the latest revision of F was and is
R2, therefore the local changes were never considered by the fragment
tracer as there do not exist any local changes based on the latest
revision R2.

The new behaviour associates local changes with R2, so we have:
  R1 < R2
  R2 < $
Consequently, local changes are visible when the fragment tracer looks
for the most recent fragments or files, respectively.