model: fix tracing of local changes in files serving as copy sources
When tracing local changes, non-local copies must be ignored as
otherwise the wrong most recent file is chosen. Note that this only
works if there is only at most one locally changed file per stop.
This is usually the case unless one makes local copies; in this case,
the fragment or file returned is chosen indeterministically.