model/svn: handling of local changes moved entirely to ChangeManager
The ChangeManager now completely manages local changes by monitoring
local files using Eclipse's resource change listener mechanism. If a
change is detected, the change source is notified using a simplified
interface (analyzeLocalChanges()) which then updates its local file
history graph. No Change or Commit objects are ever created for local
changes anymore as nobody needs them.

As a by-product of refactoring, both the list of Eclipse projects and
the reference to the change source have been moved to the ChangeManager
as well. Thus, it was possible to reduce the code complexity of the
ReviewPlugin class and to increase its cohesion.

Note that the analysis of a working copy due to adding a project to the
Subversion change source is not done in the background anymore, as it
was possible that analyzeLocalChanges() was executed before the
background job started. As the caller of addProject() already runs in
the background, however, this is not much of a problem.