model: FileHistoryGraph.findAncestorFor() generalized
This commit consists of two changes. The first one is that the operation
findAncestorFor() has been generalized to findAncestorsFor(), as a node
may have multiple ancestors (in the case of a merge). The second one is
that findAncestorsOf() has been implemented generically using the
algorithms in the PartialOrderAlgorithms class. So FileHistoryGraph
need not be subclassed by a change source anymore.