svn: ISvnRepo: added File and getFiles()
Now it is possible to compute the whole file tree for a given directory
at a given revision.

Implementation notes: This function has been implemented by using the
"svn status" machinery as opposed to using "svn list", as it is much
faster: Only one Subversion command is needed to be sent to the server
as the file tree recursion is done there and not on the client.