model: FileHistoryGraph: directory handling removed
Handling directories explicitly is only relevant for the very few
version control systems where directories are versioned objects, which,
as far as I know, only applies to Subversion and Bazaar. Much complexity
and performance problems result from supporting directory operations in
the file history graph, as the parent/child relationship needs to be
tracked and considered. However, typically this does not result in
deeper insights during code reviews. Most of the time, only the files
that are added, moved, deleted, or changed help in understanding the
commit in question. Moreover, visualizing the differences between
directories is hard (and is not done by CoRT either). So this commit
removes directory handling completely.