svn: determine children when directory is copied, deleted, or replaced
This is necessary in order to not risk defective history. This would be
not a problem if the history was complete, but we deliberately want to
handle partial history starting at an arbitrary point in time, so we
need to collect the necessary information when handling directories.

Implementation note: In Subversion, it is possible to delete a path
where one of its parent paths has been added as a copy in the same
commit. An example:

Changed paths:
[...]
   A /branches/4.0/trunk/src/packages/src/src/fbr/buildroot/linux/linux-kernel-4_4 (from /branches/4.0/trunk/src/packages/src/src/fbr/buildroot/linux/linux-kernel-4_3:44244)
[...]
   D /branches/4.0/trunk/src/packages/src/src/fbr/buildroot/linux/linux-kernel-4_4/linux-kernel-4_3_default
   D /branches/4.0/trunk/src/packages/src/src/fbr/buildroot/linux/linux-kernel-4_4/linux-kernel-4_3_nonfree
   D /branches/4.0/trunk/src/packages/src/src/fbr/buildroot/linux/linux-kernel-4_4/linux-kernel-4_3_virt
   D /branches/4.0/trunk/src/packages/src/src/fbr/buildroot/linux/linux-kernel-4_4/linux-kernel-4_3_virt_nonfree
   A /branches/4.0/trunk/src/packages/src/src/fbr/buildroot/linux/linux-kernel-4_4/linux-kernel-4_4.inc (from /branches/4.0/trunk/src/packages/src/src/fbr/buildroot/linux/linux-kernel-4_3/linux-kernel-4_3.inc:44244)
[...]

The idea behind such a commit is to copy or move a whole directory in
the hope that most of it can be reused "as is", and then to "fix" those
those parts that do not fit. Fixing can include all sorts of Subversion
operations, including changes, additions, copies, deletions, moves, or
replacements. In the case of deletions (including move and replace
operations), they have to be applied to the revision of the commit that
is being processed, not to its ancestor. In order to distinguish these
cases, paths added as a copy in the commit being analyzed are stored in
two trees (one for the paths referred by the copy operation, one for
the nested file nodes).