model/svn: repository refers again to the URL the working copy points at
In commit ff299229acf41f59377f574a1f2b119360f672ac
("model/svn: IWorkingCopy extracted from IRepository"), the code has
been changed in such a way that the Subversion repository object
created for a Subversion working copy always points at the root of the
corresponding URL, even if the working copy points at some path therein
(which is normally the case as one checks out the trunk or some
specific branch). This improves sharing; however, it also leads to much
time spent in analyzing changesets that are uninteresting to the
developer as they refer to changes in branches the developer need not
work with (and probably does not know about). This can cause a
performance penalty when branches are often created. Therefore, this
changeset restores the original situation such that a repository is
configured with the repository URL of the working copy.

Note that this also affects the computation of the latest revision:
SVNRepository.getLatestRevision() cannot be used anymore as its result
does not consider the configured path beneath the repository root.