model: FileHistoryGraph: node injection fixed
If a path is added (R1), deleted, and then added again (R2), the node
in R2 was injected before R1, as the code assumed that if an ancestor
exists, it had to be the latest one. However, this conclusion is wrong
because deletions interrupt flows, and the ancestor is the root node in
such cases. This changeset fixes the problem by adding an explicit
check.